Improvement to doodle toolbar stuff

+++ app/app/src/main/java/org/zakariya/photodoodle/fragments/DoodleFragment.java
@@ -46,6 +46,7 @@ public class DoodleFragment extends Fragment {
 	private static final String DOODLE_STATE = "DOODLE_STATE";
 
 	private static final int REQUEST_TAKE_PHOTO = 1;
+	private static final boolean DEBUG_KEEP_LAST_PHOTO = true;
 
 	enum BrushType {PENCIL, BRUSH, LARGE_ERASER, SMALL_ERASER}
 
@@ -135,6 +136,14 @@ public class DoodleFragment extends Fragment {
 			Bundle doodleState = savedInstanceState.getBundle(DOODLE_STATE);
 			if (doodleState != null) {
 				doodle.onCreate(doodleState);
+			}
+		}
+
+		if (DEBUG_KEEP_LAST_PHOTO && doodle.getPhoto() == null) {
+			Log.w(TAG, "onCreate - loading snap.jpg into doodle to simplify matrix testing");
+			Bitmap bitmap = loadPhotoFromFile(getPhotoTempFile());
+			if (bitmap != null) {
+				doodle.setPhoto(bitmap);
 			}
 		}
 
@@ -262,9 +271,14 @@ public class DoodleFragment extends Fragment {
 		}
 	}
 
-	@OnClick(R.id.cropPhotoButton)
-	void cropPhoto() {
-		Log.w(TAG, "cropPhoto - UNIMPLEMENTED");
+	@OnClick(R.id.clearDrawingButton)
+	void clearDrawing() {
+		doodle.clearDrawing();
+	}
+
+	@OnClick(R.id.clearPhotoButton)
+	void clearPhoto() {
+		doodle.clearPhoto();
 	}
 
 	public int getColor() {
@@ -395,6 +409,11 @@ public class DoodleFragment extends Fragment {
 	}
 
 	private void deletePhotoTempFile() {
+		if (DEBUG_KEEP_LAST_PHOTO) {
+			Log.w(TAG, "disabled deletePhotoTempFile while testing...");
+			return;
+		}
+
 		File photoTempFile = getPhotoTempFile();
 		if (photoTempFile.exists()) {
 			if (!photoTempFile.delete()){
